feat(Organization): Move essentialsc.h to essentialsc
