+++ src/essentialsc.c
@@ -1,4 +1,4 @@
-#include "essentialsc.h"
+#include "essentialsc/essentialsc.h"
 
 #include <stdio.h>
